Given an array of integers
Complete the plusMinus function in the editor below.

plusMinus has the following parameter(s):

int arr[n]: an array of integers
Print
Print the ratios of positive, negative and zero values in the array. Each value should be printed on a separate line with 6 digits after the decimal. The function should not return a value.

Output Format

Print the following 3 lines, each to  decimals:

1. proportion of positive values
2. proportion of negative values
3. proportion of zeros
4. Sample Input

Read the Readme for more information an examples